chore: simplify the API client
Preparation for adding support for Gems, Nuget, and Cocoapods.

Making the `SnykClient` more generic, so that it can be more easily
re-used for the REST PURL endpoint. Moving the ecosystem-specific
methods into the ecosystem scanners so that it's easier to comprehend
each scanner.
