+++ core/src/test/java/io/snyk/plugins/artifactory/SnykPluginTest.java
@@ -3,8 +3,8 @@ package io.snyk.plugins.artifactory;
 import io.snyk.plugins.artifactory.exception.SnykRuntimeException;
 import io.snyk.plugins.artifactory.util.SnykConfigForTests;
 import io.snyk.sdk.SnykConfig;
-import io.snyk.sdk.api.v1.SnykHttpRequestBuilder;
-import io.snyk.sdk.api.v1.SnykResult;
+import io.snyk.sdk.api.SnykHttpRequestBuilder;
+import io.snyk.sdk.api.SnykResult;
 import io.snyk.sdk.model.NotificationSettings;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
