+++ snyk-sdk/src/main/java/io/snyk/sdk/api/SnykHttpRequestBuilder.java
@@ -1,4 +1,4 @@
-package io.snyk.sdk.api.v1;
+package io.snyk.sdk.api;
 
 import io.snyk.sdk.SnykConfig;
